# api-spec.json
{
    "swagger": "2.0",
    "info": {
        "title": "Meet API",
        "description": "Move your app forward with the Meet API\n\n## API Success Code\n\nCode  | HTTP Code | Message\n----- | --------- | --------\n200001 | 200 | You can go to check your email\n200002 | 200 | Account validation success\n200003 | 200 | Signin success\n200004 | 200 | Create a new post successfully\n200005 | 200 | You have allowed to post articles\n200006 | 200 | Get some articles\n\n\n## API Error Code - 400\n\nError Code  | HTTP Code | Message\n------| --- | -----------------------------------\n401001 | 401 | Email or password is not valid\n401002 | 401 | Token is not valid or expired\n401003 | 401 | Email token is not valid or expired\n401004 | 401 | No Authorization header found\n401004 | 401 | Invalid Token\n403001 | 403 | The email has already been registered in social account\n403002 | 403 | The email has already been registered\n403003 | 403 | You can't repeatly post article in 3 minutes\n404001 | 404 | Your email account is not activated\n404002 | 404 | We could not find your account\n422001 | 422 | Length of password must >= 6\n\n## API Error Code - 500\n\nError Code  | HTTP Code | Message\n------| --- | -----------------------------------    \n500001 | 500 | Bcrypt error\n500002 | 500 | Your data is bad and you should feel bad\n503001 | 503 | SMTP server unavailable to verify\n503002 | 503 | SMTP server unavailable to send\n\n ## API Error Code - Bonus\n \nError Code  | HTTP Code | Message\n------| --- | -----------------------------------    \n-1 | Unknow | Unknow\n",
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/posts": {
            "post": {
                "summary": "Post a new article",
                "description": "Post a new article",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Article information",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/newPostResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden to post response",
                        "schema": {
                            "$ref": "#/definitions/forbidPostResponse"
                        }
                    },
                    "default": {
                        "description": "Error response",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/posts/findPresent": {
            "get": {
                "summary": "Get present articles",
                "description": "Get present articles",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/someArticles"
                        }
                    }
                }
            }
        },
        "/posts/findOlder": {
            "get": {
                "summary": "Get older articles",
                "description": "Get older articles",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "postID",
                        "description": "The post's objectID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/someArticles"
                        }
                    }
                }
            }
        },
        "/posts/findNewer": {
            "get": {
                "summary": "Get newer articles",
                "description": "Get newer articles",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "postID",
                        "description": "The post's objectID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/someArticles"
                        }
                    }
                }
            }
        },
        "/checkPostLimit": {
            "get": {
                "summary": "Verify the user can post articles or not",
                "description": "Verify the user can post articles or not",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/forbidPostResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden to post response",
                        "schema": {
                            "$ref": "#/definitions/forbidPostResponse"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "summary": "Signup a new account",
                "description": "Signup a new account locally",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User account information",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Signup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "default": {
                        "description": "Error response",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "get": {
                "summary": "Verify the email token",
                "description": "After the user signup a new account, ........",
                "parameters": [
                    {
                        "in": "query",
                        "name": "token",
                        "description": "email token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetUserProfileResponse"
                        }
                    },
                    "default": {
                        "description": "Error response",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/signin": {
            "post": {
                "summary": "Signin the system",
                "description": "Signin the system",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User email and password",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Signin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetUserProfileResponse"
                        }
                    },
                    "default": {
                        "description": "Error response",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/forgot_password": {
            "post": {
                "summary": "Recover password",
                "description": "The machanism for user to recover his password",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "email",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forgot_password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "default": {
                        "description": "Error response",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Signup": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "format": "email"
                },
                "password": {
                    "type": "string",
                    "format": "password"
                },
                "avatar": {
                    "type": "string",
                    "format": "byte"
                }
            }
        },
        "SuccessResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "Signin": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "format": "email"
                },
                "password": {
                    "type": "string",
                    "format": "password"
                }
            }
        },
        "Auth": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "nickname": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "format": "email"
                },
                "avatar": {
                    "type": "string",
                    "format": "byte"
                },
                "edit_nickname_time": {
                    "type": "string",
                    "format": "date"
                },
                "created_time": {
                    "type": "string",
                    "format": "date"
                },
                "updated_time": {
                    "type": "string",
                    "format": "date"
                }
            }
        },
        "newPostResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "post": {
                    "$ref": "#/definitions/Post"
                }
            }
        },
        "someArticles": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Post"
                    }
                }
            }
        },
        "Post": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "author": {
                    "$ref": "#/definitions/author"
                },
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "like_count": {
                    "type": "integer"
                },
                "dislike_count": {
                    "type": "integer"
                },
                "comment": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_time": {
                    "type": "string",
                    "format": "date"
                },
                "updated_time": {
                    "type": "string",
                    "format": "date"
                }
            }
        },
        "author": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "nickname": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "format": "email"
                },
                "avatar": {
                    "type": "string",
                    "format": "byte"
                },
                "edit_nickname_time": {
                    "type": "string",
                    "format": "date"
                },
                "created_time": {
                    "type": "string",
                    "format": "date"
                },
                "updated_time": {
                    "type": "string",
                    "format": "date"
                }
            }
        },
        "GetUserProfileResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "auth": {
                    "$ref": "#/definitions/Auth"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "forgot_password": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "format": "email"
                }
            }
        },
        "PostRequest": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                }
            }
        },
        "forbidPostResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "create_post_time": {
                    "type": "string",
                    "format": "date"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}